#1 Refactor authentication
