+++ Core/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="RestSharp" version="105.2.3" targetFramework="net452" />
+</packages>
